+++ lib/providers/accelerometer_config_provider.dart
@@ -58,7 +58,7 @@ class AccelerometerConfigProvider extends ChangeNotifier {
     _saveConfigToPrefs();
   }
 
-  void updateSensorGain(int sensorGain) {
+  void updateSensorGain(double sensorGain) {
     _config = _config.copyWith(sensorGain: sensorGain);
     notifyListeners();
     _saveConfigToPrefs();
